+++ packages/switch/src/css/index.js
@@ -2,8 +2,8 @@ import core from '@pluralsight/ps-design-system-core'
 import {
   defaultName as themeDefaultName,
   names as themeNames
-} from '@pluralsight/ps-design-system-theme/react'
-import { transparentize } from '@pluralsight/ps-design-system-util/color'
+} from '@pluralsight/ps-design-system-theme/react.js'
+import { transparentize } from '@pluralsight/ps-design-system-util/color.js'
 
 import * as vars from '../vars/index.js'
 
@@ -15,10 +15,9 @@ export default {
     opacity: '1',
     background: 'none',
     border: 'none',
-    padding: 0
-  },
-  '.psds-switch:focus': {
-    outline: 'none'
+    padding: 0,
+
+    '&:focus': { outline: 'none' }
   },
   '.psds-switch--disabled': {
     opacity: '0.4',
@@ -47,19 +46,13 @@ export default {
     borderRadius: '12px',
     padding: '1px'
   },
-  [`.psds-switch__track--checked.psds-switch__track--color-${
-    vars.colors.blue
-  }`]: {
+  [`.psds-switch__track--checked.psds-switch__track--color-${vars.colors.blue}`]: {
     backgroundColor: core.colors.blue
   },
-  [`.psds-switch__track--checked.psds-switch__track--color-${
-    vars.colors.green
-  }`]: {
+  [`.psds-switch__track--checked.psds-switch__track--color-${vars.colors.green}`]: {
     backgroundColor: core.colors.green
   },
-  [`.psds-switch__track--checked.psds-switch__track--color-${
-    vars.colors.orange
-  }`]: {
+  [`.psds-switch__track--checked.psds-switch__track--color-${vars.colors.orange}`]: {
     backgroundColor: core.colors.orange
   },
 
@@ -77,14 +70,10 @@ export default {
     height: '22px',
     width: '22px'
   },
-  [`.psds-switch__thumb--checked.psds-switch__thumb--size-${
-    vars.sizes.small
-  }`]: {
+  [`.psds-switch__thumb--checked.psds-switch__thumb--size-${vars.sizes.small}`]: {
     transform: 'translateX(12px)'
   },
-  [`.psds-switch__thumb--checked.psds-switch__thumb--size-${
-    vars.sizes.large
-  }`]: {
+  [`.psds-switch__thumb--checked.psds-switch__thumb--size-${vars.sizes.large}`]: {
     transform: 'translateX(24px)'
   },
 
@@ -98,24 +87,16 @@ export default {
   [`.psds-switch__label--size-${vars.sizes.large}`]: {
     fontSize: core.type.fontSizeSmall
   },
-  [`.psds-switch__label--size-${
-    vars.sizes.small
-  }.psds-switch__label--labelAlign-${vars.labelAligns.left}`]: {
+  [`.psds-switch__label--size-${vars.sizes.small}.psds-switch__label--labelAlign-${vars.labelAligns.left}`]: {
     marginRight: core.layout.spacingXSmall
   },
-  [`.psds-switch__label--size-${
-    vars.sizes.large
-  }.psds-switch__label--labelAlign-${vars.labelAligns.left}`]: {
+  [`.psds-switch__label--size-${vars.sizes.large}.psds-switch__label--labelAlign-${vars.labelAligns.left}`]: {
     marginRight: core.layout.spacingMedium
   },
-  [`.psds-switch__label--size-${
-    vars.sizes.small
-  }.psds-switch__label--labelAlign-${vars.labelAligns.right}`]: {
+  [`.psds-switch__label--size-${vars.sizes.small}.psds-switch__label--labelAlign-${vars.labelAligns.right}`]: {
     marginLeft: core.layout.spacingXSmall
   },
-  [`.psds-switch__label--size-${
-    vars.sizes.large
-  }.psds-switch__label--labelAlign-${vars.labelAligns.right}`]: {
+  [`.psds-switch__label--size-${vars.sizes.large}.psds-switch__label--labelAlign-${vars.labelAligns.right}`]: {
     marginLeft: core.layout.spacingMedium
   },
   [`.psds-switch__label.psds-theme--${themeNames.light}`]: {