+++ packages/textarea/src/css/index.js
@@ -5,7 +5,8 @@ import { names as themeNames } from '@pluralsight/ps-design-system-theme/vars'
 export default {
   '.psds-text-area': {
     display: 'inline-block',
-    width: '100%'
+    width: '100%',
+    maxWidth: '496px'
   },
   '.psds-text-area--disabled': {
     opacity: 0.5
@@ -41,10 +42,7 @@ export default {
   '.psds-text-area__field-container': {
     position: 'relative',
     display: 'flex',
-    alignItems: 'center',
-    minWidth: `calc(192px + ${iconVars.widths.medium} + ${
-      core.layout.spacingXSmall
-    })`
+    alignItems: 'center'
   },
 
   '.psds-text-area__label': {
@@ -72,23 +70,12 @@ export default {
 
   '.psds-text-area__error': {
     position: 'absolute',
-    right: `calc(-1 * (${iconVars.widths.medium} + ${
-      core.layout.spacingXSmall
-    }))`,
+    right: `calc(-1 * (${iconVars.widths.medium} + ${core.layout.spacingXSmall}))`,
     display: 'flex',
     alignItems: 'flex-start',
     height: '100%',
     paddingTop: core.layout.spacingXSmall,
     marginLeft: core.layout.spacingXSmall,
     color: core.colors.red
-  },
-
-  '@media (min-width: 433px)': {
-    '.psds-text-area': {
-      width: 'auto'
-    },
-    '.psds-text-area__field': {
-      minWidth: '433px'
-    }
   }
 }