+++ packages/site/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [13.2.1](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@13.2.0...@pluralsight/ps-design-system-site@13.2.1) (2019-09-05)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
+# [13.2.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@13.1.5...@pluralsight/ps-design-system-site@13.2.0) (2019-09-05)
+
+
+### Bug Fixes
+
+* **card:** adds tag to small card for issue ([c4b0225](https://github.com/pluralsight/design-system/commit/c4b0225)), closes [#549](https://github.com/pluralsight/design-system/issues/549) [#551](https://github.com/pluralsight/design-system/issues/551)
+
+
+### Features
+
+* **card:** derive aria-label from title prop ([d158446](https://github.com/pluralsight/design-system/commit/d158446))
+
+
+
+
+
+## [13.1.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@13.1.4...@pluralsight/ps-design-system-site@13.1.5) (2019-08-30)
+
+
+### Bug Fixes
+
+* **site:** restore http-server needed for runtime ([a71ba38](https://github.com/pluralsight/design-system/commit/a71ba38))
+
+
+
+
+
+## [13.1.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@13.1.3...@pluralsight/ps-design-system-site@13.1.4) (2019-08-28)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
+## [13.1.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@13.1.2...@pluralsight/ps-design-system-site@13.1.3) (2019-08-28)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-site
+
+
+
+
+
 ## [13.1.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-site@13.1.1...@pluralsight/ps-design-system-site@13.1.2) (2019-08-27)